requirements.txt
lxml
cssselect
requests
python-dotenv
